add model and bg blur
