+++ CMakeLists.txt
@@ -4,7 +4,7 @@ set(CMAKE_CXX_STANDARD 20)
 
 # Change obs-plugintemplate to your plugin's name in a machine-readable format (e.g.:
 # obs-myawesomeplugin) and set
-project(obs-backgroundremoval VERSION 0.5.2)
+project(obs-backgroundremoval VERSION 0.5.3)
 add_library(${CMAKE_PROJECT_NAME} MODULE)
 
 # Replace `Your Name Here` with the name (yours or your organization's) you want to see as the
