+++ buildspec.json
@@ -82,5 +82,5 @@
         }
     },
     "name": "obs-backgroundremoval",
-    "version": "0.5.2"
+    "version": "0.5.3"
 }
